initialized jasmine unittests
