+++ jasmine.json
@@ -1,0 +1,6 @@
+{
+    "spec_dir": "./tests",
+    "spec_files": [ "**/*.spec.ts" ],
+    "stopSpecOnExpectationFailure": true,
+    "random": true
+}
