+++ tests/index.spec.ts
@@ -1,0 +1,7 @@
+import { run } from "../src/index";
+
+describe("test index file", () => {
+    it("should exist", () => {
+        expect(run).toBeTruthy();
+    });
+});
