+++ tests/utils.spec.ts
@@ -1,0 +1,7 @@
+import { addReport } from "../src/utils";
+
+describe("test addReport method", () => {
+    it("should exist", () => {
+        expect(addReport).toBeTruthy();
+    });
+})
