+++ tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "target": "es2015",
     "moduleResolution": "node",
+    "sourceMap": true
   },
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "**/*.spec.ts"]
 }
